requirements.txt
matplotlib
numpy
pandas
scikit-learn
wandb
torchvision
torch
requests
python-dotenv
ray[default]